TST: Include test players as fixtures.
