+++ test/testplayers/remote_dumps_with_failure_bad.py
@@ -1,0 +1,6 @@
+TEAM_NAME="failing"
+def move(b, s):
+    if b.round == 2 and b.turn == 0:
+        # introduce an error
+        0 / 0
+    return b.position
