+++ test/testplayers/remote_dumps_with_failure_good.py
@@ -1,0 +1,3 @@
+TEAM_NAME="good"
+def move(b, s):
+    return b.position
